requirements.txt
numpy>=1.20.0
scipy>=1.7.0
pyyaml>=6.0
pytest>=7.0.0
